+++ cnext_app/__init__.py
@@ -1,2 +1,0 @@
-#for control version in code
-__version__ = '0.4.22'