+++ cnext_app/lib/components/richoutput-panel/data-panel/DFFilter.tsx
@@ -1,10 +1,12 @@
 import { cnextQuery } from "../../../codemirror/grammar/lang-cnext-query";
-import React, { useRef } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { useDispatch } from "react-redux";
-import { DFFilterForm, DFFilterInput, StyledFilterCodeMirror } from "../../StyledComponents";
-// import { bracketMatching } from "@codemirror/matchbrackets";
-// import { closeBrackets } from "@codemirror/closebrackets";
-// import { defaultHighlightStyle } from "@codemirror/highlight";
+import {
+    DFFilterForm,
+    DFFilterInput,
+    QuerySample,
+    StyledFilterCodeMirror,
+} from "../../StyledComponents";
 import { dfFilterLanguageServer } from "../../../codemirror/autocomplete-lsp/index.js";
 import { setDFFilter } from "../../../../redux/reducers/DataFramesRedux";
 import store from "../../../../redux/store";
@@ -19,10 +21,47 @@ import { setRichOutputFocused } from "../../../../redux/reducers/RichOutputRedux
 
 const ls = dfFilterLanguageServer();
 
-const DFFilter = React.memo(() => {
-    // const dfList = useSelector((state) => _checkDFList(state));
+interface DFQuery {
+    df_id: string;
+    query: string | null;
+    /** String that is entered into the query box. */
+    raw_query: string;
+}
+
+const DFExplorer = () => {
+    const [query, setQuery] = useState<DFQuery | null>(null);
     const dispatch = useDispatch();
     const filterCM = useRef();
+
+    const keyHandler = useCallback(
+        (event: React.KeyboardEvent) => {
+            // console.log("DFExplorer event: ", event);
+            try {
+                if (event.key === "Enter" && query != null) {
+                    console.log(
+                        `DFExplorer dispatch query: ${query}, codeEditorText: ${codeEditorText}`
+                    );
+                    let codeEditorText =
+                        filterCM?.current?.getElementsByClassName("cm-line")[0].innerText;
+                    /** use this trick to avoid the enter keyboard event that is triggered by autocompletion  */
+                    if (query.raw_query === codeEditorText) {
+                        dispatch(setDFFilter(query));
+                    }
+                }
+            } catch {}
+        },
+        [query]
+    );
+
+    useEffect(() => {
+        let element = filterCM.current;
+        if (element != null) {
+            element.addEventListener("keydown", keyHandler);
+            return () => {
+                element.removeEventListener("keydown", keyHandler);
+            };
+        }
+    }, [keyHandler]);
 
     /**
      * All query string will be converted to loc/iloc pandas query
@@ -38,11 +77,10 @@ const DFFilter = React.memo(() => {
                 tree = v.context.tree;
             }
         }
-        let queryStr;
+        let queryStr = null;
         //[1:2, ['a']]['a'>2]
         //[1:2, ['a']][('a'>2) & ('b'<4) | ('c'>5)]
         if (tree) {
-            console.log(tree.toString());
             let cursor = tree.cursor(0, 0);
             let curComponent;
             if (cursor.name == "Script" && cursor.firstChild() && cursor.name == "QueryStatement") {
@@ -126,7 +164,10 @@ const DFFilter = React.memo(() => {
                 }
             }
         }
-        dispatch(setDFFilter({ df_id: activeDF, query: queryStr }));
+        if (activeDF != null) {
+            console.log("DFExplorer query: ", queryStr);
+            setQuery({ df_id: activeDF, query: queryStr, raw_query: text });
+        }
     };
 
     const oneLineExtension = () => {
@@ -139,7 +180,7 @@ const DFFilter = React.memo(() => {
                 return null;
             }
             return transaction;
-        })
+        });
     };
 
     const extensions = [
@@ -161,16 +202,16 @@ const DFFilter = React.memo(() => {
             return transaction;
         }),
         oneLineExtension(),
+        // keymap.of([{ key: "Enter", run: () => enterKeyHandler() }]),
     ];
 
     return (
         <DFFilterForm>
             <DFFilterInput
-                placeholder="Filter..."
-                inputComponent={() => {
+                ref={filterCM}
+                inputComponent={useCallback(() => {
                     return (
                         <StyledFilterCodeMirror
-                            ref={filterCM}
                             extensions={extensions}
                             basicSetup={false}
                             onChange={(text, viewUpdate) => onCMChange(text, viewUpdate)}
@@ -182,11 +223,11 @@ const DFFilter = React.memo(() => {
                             }}
                         />
                     );
-                }}
-                
+                }, [])}
             ></DFFilterInput>
+            {query && query.query && <QuerySample>{query.df_id + query.query}</QuerySample>}
         </DFFilterForm>
     );
-});
+};
 
-export default DFFilter;
+export default DFExplorer;