+++ cnext_app/lib/components/richoutput-panel/result-panel/ResultContent.tsx
@@ -20,11 +20,12 @@ const ScriptComponent = ({ children, script }) => {
         var scriptElem = document.createElement("script");
         // console.log("ResultContent script ", script);
         if (script != null) {
-            if (script.attribs != null && script.attribs.src != null)
+            if (script.attribs != null && script.attribs.src != null) {
                 scriptElem.src = script.attribs.src;
+            }
             if (script.children != null) {
                 for (let node of script.children) {
-                    var textNode = document.createTextNode(node.data);
+                    let textNode = document.createTextNode(node.data);
                     scriptElem.appendChild(textNode);
                 }
             }
@@ -42,20 +43,27 @@ const ScriptComponent = ({ children, script }) => {
     return <div ref={instance}></div>;
 };
 
-const IFrameComponent = ({ children }) => {
+const IFrameComponent = ({ children, attribs, stopMouseEvent }) => {
     const [contentRef, setContentRef] = useState(null);
     const mountNode = contentRef?.contentDocument?.body;
     return (
         <iframe
-            style={{ width: "100%", height: "100vh", border: "0px solid white" }}
+            style={{
+                width: "100%",
+                height: "100vh",
+                border: "0px solid white",
+                padding: "0px",
+                pointerEvents: stopMouseEvent?"none":"auto",
+            }}
             ref={setContentRef}
+            srcDoc={attribs?.srcdoc}
         >
-            {mountNode && createPortal(children, mountNode)}
+            {children && mountNode && createPortal(children, mountNode)}
         </iframe>
     );
 };
 
-const ResultContent = React.memo(({ codeResult }) => {
+const ResultContent = React.memo(({ codeResult, showMarkdown, stopMouseEvent }) => {
     // const [readyToScroll, setReadyToScroll] = useState(false);
 
     const setPlotlyLayout = (
@@ -87,87 +95,116 @@ const ResultContent = React.memo(({ codeResult }) => {
     };
 
     const renderResultContent = () => {
-        // const imageMime = getMimeWithImage(Object.keys(codeResult?.result?.content));
-        // console.log("ResultContent: ", codeResult?.result);
-        const showMarkdown = store.getState().projectManager.settings?.rich_output?.show_markdown;
-        const contentKeys = Object.keys(codeResult?.result?.content);
-        // console.log("ResultContent contentKeys: ", contentKeys);
-        let jsxElements = contentKeys?.map((key, index) => {
-            const imageMime = getMimeWithImage([key]);
-            if (key === SubContentType.APPLICATION_JAVASCRIPT) {
-                return (
-                    <ScriptComponent script={null}>
-                        {codeResult?.result?.content[
-                            SubContentType.APPLICATION_JAVASCRIPT
-                        ].toString("base64")}
-                    </ScriptComponent>
-                );
-            } else if (key === SubContentType.APPLICATION_BOKEH) {
-                return (
-                    <ScriptComponent script={null}>
-                        {codeResult?.result?.content[SubContentType.APPLICATION_BOKEH].toString(
-                            "base64"
-                        )}
-                    </ScriptComponent>
-                );
-            } else if (key === SubContentType.APPLICATION_PLOTLY) {
-                return React.createElement(
-                    PlotlyWithNoSSR,
-                    setPlotlyLayout(codeResult?.result?.content[SubContentType.APPLICATION_PLOTLY])
-                );
-            } else if (key === SubContentType.APPLICATION_JSON) {
-                return JSON.stringify(codeResult?.result?.content[SubContentType.APPLICATION_JSON]);
-            } else if (imageMime != null) {
-                // console.log("ResultView ", imageMime, codeResult?.result?.content[imageMime]);
-                return (
-                    <img
-                        src={
-                            "data:" +
-                            imageMime +
-                            ";base64," +
-                            codeResult?.result?.content[imageMime]
-                        }
-                    />
-                );
-            } else if (key === SubContentType.TEXT_HTML) {
-                const htmlRegex = new RegExp("<!DOCTYPE html>|<html>");
-                const htmlContent = codeResult?.result?.content[SubContentType.TEXT_HTML];
-                // console.log("ResultContent text/html content: ", htmlContent);
-                let isFullPage = htmlRegex.test(htmlContent);
-                let jsxElements = ReactHtmlParser(htmlContent.toString("base64"), {
-                    replace: function (domNode) {
-                        // console.log("ResultContent domNode ", domNode);
-                        if (domNode.type === "script") {
-                            return <ScriptComponent children={null} script={domNode} />;
-                        }
-                    },
-                });
+        try {
+            // const imageMime = getMimeWithImage(Object.keys(codeResult?.result?.content));
+            // console.log("ResultContent: ", codeResult?.result);
+            // const showMarkdown = store.getState().projectManager.settings?.rich_output?.show_markdown;
+            const contentKeys = Object.keys(codeResult?.result?.content);
+            // console.log("ResultContent contentKeys: ", contentKeys);
+            let jsxElements = contentKeys?.map((key, index) => {
+                const imageMime = getMimeWithImage([key]);
+                if (key === SubContentType.APPLICATION_JAVASCRIPT) {
+                    return (
+                        <ScriptComponent script={null}>
+                            {codeResult?.result?.content[
+                                SubContentType.APPLICATION_JAVASCRIPT
+                            ].toString("base64")}
+                        </ScriptComponent>
+                    );
+                } else if (key === SubContentType.APPLICATION_BOKEH) {
+                    return (
+                        <ScriptComponent script={null}>
+                            {codeResult?.result?.content[SubContentType.APPLICATION_BOKEH].toString(
+                                "base64"
+                            )}
+                        </ScriptComponent>
+                    );
+                } else if (key === SubContentType.APPLICATION_PLOTLY) {
+                    return React.createElement(
+                        PlotlyWithNoSSR,
+                        setPlotlyLayout(
+                            codeResult?.result?.content[SubContentType.APPLICATION_PLOTLY]
+                        )
+                    );
+                } else if (key === SubContentType.APPLICATION_JSON) {
+                    return JSON.stringify(
+                        codeResult?.result?.content[SubContentType.APPLICATION_JSON]
+                    );
+                } else if (imageMime != null) {
+                    // console.log("ResultView ", imageMime, codeResult?.result?.content[imageMime]);
+                    return (
+                        <img
+                            src={
+                                "data:" +
+                                imageMime +
+                                ";base64," +
+                                codeResult?.result?.content[imageMime]
+                            }
+                        />
+                    );
+                } else if (key === SubContentType.TEXT_HTML) {
+                    const htmlRegex = new RegExp("<!DOCTYPE html|<html", "i");
+                    const iframeRegex = new RegExp("<iframe", "i");
+                    const htmlContent = codeResult?.result?.content[SubContentType.TEXT_HTML];
+                    // console.log("ResultContent text/html content: ", htmlContent);
+                    let isIFrame = iframeRegex.test(htmlContent);
+                    let isHTMLPage = htmlRegex.test(htmlContent);
+                    let jsxElements = ReactHtmlParser(htmlContent.toString("base64"), {
+                        replace: function (domNode) {
+                            // console.log("ResultContent domNode ", domNode);
+                            if (domNode.type === "script") {
+                                return <ScriptComponent children={null} script={domNode} />;
+                            } else if (domNode.type === "tag" && domNode.name === "iframe") {
+                                return (
+                                    <IFrameComponent
+                                        attribs={domNode.attribs}
+                                        children={domNode.children}
+                                        stopMouseEvent={stopMouseEvent}
+                                    ></IFrameComponent>
+                                );
+                            }
+                        },
+                    });
 
-                return isFullPage ? (
-                    <IFrameComponent children={jsxElements}></IFrameComponent>
-                ) : (
-                    jsxElements
-                );
-            } else if (showMarkdown && key === SubContentType.MARKDOWN) {
-                return (
-                    <ReactMarkdown
-                        className="markdown"
-                        remarkPlugins={[remarkGfm]}
-                        children={codeResult?.result?.content[SubContentType.MARKDOWN]}
-                    />
-                );
-            } else if (key === SubContentType.TEXT_PLAIN && contentKeys.length === 1) {
-            /** only display text/plain when it is the only content */
-                return (
-                    <pre
-                        style={{ fontSize: "12px" }}
-                        children={codeResult?.result?.content[SubContentType.TEXT_PLAIN]}
-                    />
-                );
-            }
-        });
-        console.log("ResultContent renderResultContent: ", jsxElements);
-        return <div>{jsxElements}</div>;
+                    if (isIFrame && jsxElements != null) {
+                        return jsxElements;
+                    } else if (isHTMLPage) {
+                        return (
+                            <IFrameComponent
+                                children={jsxElements}
+                                stopMouseEvent={stopMouseEvent}
+                            ></IFrameComponent>
+                        );
+                    } else {
+                        return jsxElements;
+                    }
+                } else if (
+                    showMarkdown &&
+                    key === SubContentType.MARKDOWN &&
+                    codeResult?.result?.content[SubContentType.MARKDOWN] != null
+                ) {
+                    return (
+                        <ReactMarkdown
+                            className="markdown"
+                            remarkPlugins={[remarkGfm]}
+                            children={codeResult?.result?.content[SubContentType.MARKDOWN]}
+                        />
+                    );
+                } else if (key === SubContentType.TEXT_PLAIN && contentKeys.length === 1) {
+                    /** only display text/plain when it is the only content */
+                    return (
+                        <pre
+                            style={{ fontSize: "12px" }}
+                            children={codeResult?.result?.content[SubContentType.TEXT_PLAIN]}
+                        />
+                    );
+                }
+            });
+            console.log("ResultContent renderResultContent: ", jsxElements);
+            return <div>{jsxElements}</div>;
+        } catch (error) {
+            console.error(error);
+        }
     };
 
     // useEffect(() => {