+++ ReleaseNotes.md
@@ -1,17 +1,21 @@
-# v0.4.20 - 0.4.22
+# v0.4.20 - 0.4.29
 
 ##### NEW
 
 -   Implement add project feature
+-   Add turn on/off markdown
 
 ##### IMPROVED
 
 -   Improve setup script
 -   Improve usability
+-   Improve file explorer style
 
 ##### FIXED
 
 -   Bug fixes and usability improvement
+-   Fix bugs with panel resizing
+-   Fix file path issues on Windows system
 
 # v0.4.18 - 0.4.19